Web: move svg from toast to public path
