+++ packages/components/snackbar/styled-snackbar-logo.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import InfoIcon from "../../../public/images/alert.react.svg";
+import InfoIcon from "../../../public/images/danger.toast.react.svg";
 import commonIconsStyles from "../utils/common-icons-style";
 
 const StyledLogoIcon = styled(InfoIcon)`
