--- packages/components/toast/svg/index.js
@@ -1,3 +1,0 @@
-export { default as CheckToastIcon } from "./check.toast.react.svg";
-export { default as DangerToastIcon } from "./danger.toast.react.svg";
-export { default as InfoToastIcon } from "./info.toast.react.svg";
